requirements.txt
alembic==1.13.1
bertopic==0.16.0
Flask==3.0.2
Flask_Mail==0.9.1
Flask_Migrate==4.0.5
flask_sqlalchemy==3.1.1
hdbscan==0.8.33
pandas==2.2.0
scikit_learn==1.4.1.post1
SQLAlchemy==2.0.27
torch==2.2.0
torch_xla==2.2.0
tqdm==4.66.2
transformers==4.37.2
umap_learn==0.5.5
Werkzeug==3.0.1
WTForms==3.1.2
